# api/openapi/stackkits-v1.yaml
openapi: "3.1.0"
info:
  title: kombify StackKits API
  version: "1.0.0"
  description: |
    kombify StackKits HTTP API — pre-packaged homelab infrastructure templates
    with CUE schema validation and OpenTofu generation. Designed for consumption
    by the kombify Stack control plane and AI agents via the kombify Gateway.
  contact:
    name: kombify
    url: https://kombify.io
  license:
    name: MIT
servers:
  - url: http://localhost:8082
    description: Local development
  - url: https://api.kombify.io/stackkits
    description: Production (via Cloudflare Edge)
paths:
...
  /api/v2/resolve:
    post:
      operationId: resolveArchitectureV2
      summary: Resolve canonical Architecture v2 intent
      description: |
        Resolves a canonical StackSpec v2 and optional observed inventory through
        the embedded governed CUE authority. The successful response is the raw,
        canonical ResolvedPlan document; it is not wrapped in a success envelope.
        StackSpec v1 input is never compiled implicitly and instead returns a
        machine-readable migration report. Production deployments require the
        X-API-Key security scheme; explicitly unauthenticated local-development
        deployments may use the anonymous alternative. Requests are bounded to
        64 JSON nesting levels, 65,536 JSON tokens, and 4,096 entries per
        individual object or array before governed CUE resolution begins.
      tags: [Architecture]
      security:
        - StackKitsAPIKey: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchitectureV2ResolveRequest"
      responses:
        "200":
          description: Canonical ResolvedPlan bytes
          headers:
            Cache-Control:
              description: Resolved deployment plans must not be cached.
              schema:
                type: string
                const: no-store
            ETag:
              description: Strong ETag containing the quoted ResolvedPlan planHash.
              schema:
                type: string
                pattern: '^"sha256:[a-f0-9]{64}"$'
            X-StackKit-Plan-Hash:
              description: Unquoted ResolvedPlan planHash.
              schema:
                $ref: "#/components/schemas/ArchitectureV2ContentHash"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureV2ResolvedPlan"
        "400":
          description: Invalid JSON envelope, StackSpec, or inventory
          headers:
            Cache-Control:
              $ref: "#/components/headers/ArchitectureV2NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureV2ResolveError"
        "401":
          description: Missing API key when authentication is enabled
          headers:
            Cache-Control:
              $ref: "#/components/headers/ArchitectureV2NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid API key when authentication is enabled
          headers:
            Cache-Control:
              $ref: "#/components/headers/ArchitectureV2NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Resolve request exceeds the one-mebibyte body limit
          headers:
            Cache-Control:
              $ref: "#/components/headers/ArchitectureV2NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureV2ResolveError"
        "415":
          description: Request Content-Type is not application/json
          headers:
            Cache-Control:
              $ref: "#/components/headers/ArchitectureV2NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureV2ResolveError"
        "422":
          description: Migration is required or blocked, or governed resolution failed
          headers:
            Cache-Control:
              $ref: "#/components/headers/ArchitectureV2NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureV2ResolveError"
        "429":
          description: Per-server Architecture resolver concurrency or global request rate limit reached
          headers:
            Cache-Control:
              $ref: "#/components/headers/ArchitectureV2NoStore"
            Retry-After:
              $ref: "#/components/headers/ArchitectureV2RetryAfter"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ArchitectureV2ResolveError"
                  - $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Embedded Architecture v2 authority is unavailable
          headers:
            Cache-Control:
              $ref: "#/components/headers/ArchitectureV2NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureV2ResolveError"
        "500":
          description: Unexpected resolver or server failure
          headers:
            Cache-Control:
              $ref: "#/components/headers/ArchitectureV2NoStore"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ArchitectureV2ResolveError"
                  - $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/stackkits/{name}:
    get:
      operationId: getStackKit
      summary: Get full StackKit details
      tags: [Catalog]
      parameters:
        - $ref: "#/components/parameters/StackKitName"
      responses:
        "200":
          description: Full StackKit configuration
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - properties:
                      data:
                        $ref: "#/components/schemas/StackKit"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/stackkits/{name}/schema:
    get:
      operationId: getStackKitSchema
      summary: Get raw CUE schema for a StackKit
      tags: [Catalog]
      parameters:
        - $ref: "#/components/parameters/StackKitName"
      responses:
        "200":
          description: CUE schema content
          content:
            text/plain:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/stackkits/{name}/defaults:
    get:
      operationId: getStackKitDefaults
      summary: Read versioned initial StackSpec authoring data for a StackKit
      description: >-
        Exact v0.6 servers return the legacy StackSpec defaults. Native v0.7 servers return CUE-governed Architecture v2 authoring metadata and include a canonical initial StackSpec only when no user override is required. The v2 response is spec-only evidence and never asserts inventory, plan, generation, or execution readiness.
      tags: [Catalog]
      parameters:
        - $ref: "#/components/parameters/StackKitName"
      responses:
        "200":
          description: Default spec values
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - properties:
                      data:
                        oneOf:
                          - $ref: "#/components/schemas/StackSpec"
                          - $ref: "#/components/schemas/ArchitectureV2InitialAuthoringResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/validate:
    post:
      operationId: validateSpec
      summary: Validate a stack-spec against its StackKit schema
      description: >-
        Canonical StackSpec v2 is validated against the embedded CUE authority and produces spec-only evidence. On native v0.7, StackSpec v1 remains accepted only for read-only validation and migration diagnostics; it is never operationally admitted.
      tags: [Validation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/ArchitectureV2StackSpec"
                - $ref: "#/components/schemas/ArchitectureV1DiagnosticsInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - properties:
                      data:
                        oneOf:
                          - $ref: "#/components/schemas/ArchitectureV2SpecValidation"
                          - $ref: "#/components/schemas/ArchitectureV1ReadOnlyValidation"
                          - $ref: "#/components/schemas/LegacyValidationResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/validate/partial:
    post:
      operationId: validatePartial
      summary: Validate versioned initial StackSpec authoring input
      description: >-
        Exact v0.6 servers retain legacy partial StackSpec validation. Native v0.7 servers accept only the governed Architecture v2 initial-authoring request and materialize a complete CUE-valid StackSpec. Arbitrary partial StackSpec overlays are not accepted on the v0.7 line.
      tags: [Validation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  description: Legacy v0.6 partial StackSpec fields without an Architecture v2 discriminator.
                  not:
                    required: [apiVersion]
                - $ref: "#/components/schemas/ArchitectureV2InitialAuthoringRequest"
      responses:
        "200":
          description: Partial validation result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - properties:
                      data:
                        oneOf:
                          - $ref: "#/components/schemas/ValidationResult"
                          - $ref: "#/components/schemas/ArchitectureV2InitialAuthoringValidation"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/generate/tfvars:
    post:
      operationId: generateTFVars
      summary: Generate terraform.tfvars from a validated spec
      description: >-
        Exact-v0.6 compatibility endpoint only. Native v0.7 does not advertise this capability and returns a typed migration/not-supported error before composition. Architecture v2 generation consumes an Inventory-bound ResolvedPlan through the governed CLI/executor path.
      deprecated: true
      tags: [Generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          description: Generated tfvars
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - properties:
                      data:
                        $ref: "#/components/schemas/GenerateResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/generate/preview:
    post:
      operationId: generatePreview
      summary: Preview generated infrastructure without writing files
      description: >-
        Exact-v0.6 compatibility endpoint only. Native v0.7 does not advertise this capability and never treats spec-only input as generation-ready.
      deprecated: true
      tags: [Generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          description: Preview of generated output
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - properties:
                      data:
                        $ref: "#/components/schemas/PreviewResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /api/v1/verify:
    post:
      operationId: runManagementVerify
      summary: Run node-local StackKit verification
      description: >-
        Exact-v0.6 compatibility verifier. Native v0.7 returns 501 until this endpoint consumes the exact verified ResolvedPlan and execution evidence; use the Architecture v2 CLI/executor verification path meanwhile.
      deprecated: true
      tags: [Management]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRequest"
      responses:
        "200":
          description: Verification report
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - properties:
                      data:
                        $ref: "#/components/schemas/VerifyReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "501":
          $ref: "#/components/responses/OperationalUnavailable"
...
components:
  securitySchemes:
    StackKitsAPIKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Required by production deployments; may be disabled only for explicit local development.
  headers:
    ArchitectureV2NoStore:
      description: Architecture intent, plans, and diagnostics must not be cached.
      schema:
        type: string
        const: no-store
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: integer
            message:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ArchitectureV2ResolveRequest:
      type: object
      description: Strict HTTP envelope for Architecture v2 resolution.
      additionalProperties: false
      required: [stackSpec]
      properties:
        stackSpec:
          oneOf:
            - $ref: "#/components/schemas/ArchitectureV2StackSpec"
            - $ref: "#/components/schemas/ArchitectureV1DiagnosticsInput"
        inventory:
          $ref: "#/components/schemas/ArchitectureV2Inventory"
...
    ArchitectureV2ContentHash:
      type: string
      pattern: "^sha256:[a-f0-9]{64}$"
...